pyproject.toml
[project]
name = "query-expansion"
version = "0.1.0"
description = "Query expansion system using LLMs and knowledge graphs for RAG evaluation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "spacy>=3.8.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=0.1.0",
    "langchain-google-vertexai>=2.0.0",
    "google-genai>=1.3.0",
    "google-generativeai>=0.8.0",
    "google-cloud-aiplatform>=1.79.0",
    "pandas>=2.2.0",
    "networkx>=3.4.0",
    "matplotlib>=3.10.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.24.0",
    "coverage>=7.6.0",
]
dev = [
    "ipython>=8.32.0",
    "jupyter>=1.0.0",
    "ruff>=0.3.0",
    "pip",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["analyzer", "config", "goldendataset", "knowledgegraph", "querygenerator", "rag_evaluator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --tb=short"

[tool.coverage.run]
source = ["."]
omit = ["*/tests/*", "*/venv*/*", "*/__pycache__/*", "*/old_files/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
